host/requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
coverage==7.2.5
Django==4.2
django-cors-headers==4.0.0
djangorestframework==3.14.0
passlib==1.7.4
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
